New notebooks
Example notebooks for showing how to use the SNR computation, how to find the limiting magnitude. Sensitivity vs Efficiency has an example of how to get at the required PSF for a given *effective area* (aperture * optical efficiency. Needs to be re-run in a couple of bands for completeness, etc.