Added 1D background estimation
For the small postage stamp images that we're using for the simulations at the moment it's better to use a 1D sigmaclipped estimate of the background, as the 2D backgrounds introduce artifacts. I've added a method keyword to estimate_background that defaults to 2D (current behavior), but should be set to 1D when using estimate_background for background subtraction in small simulated images.