Remove off-image detections in 'find'
Ran into a bug where 'find' with method=daophot detects 'sources' that are off the edge of the image, which then causes run_daophot to crash because it doesn't have enough pixels to fit a psf to. This fix catches those 'detections' within find and removes them before writing out the results table.
Also added some background calculations to the OII suppression notebook.